+++ src/routes/web.php
@@ -53,6 +53,7 @@ Route::middleware('auth')->group(function () {
     Route::post('/create-comment', 'CommentController@create')->name('create-comment');
     Route::post('/delete-comment', 'CommentController@delete')->name('delete-comment');
     Route::post('/update-profile', 'UserController@update')->name('update-profile');
+    Route::post('/update-password', 'UserController@updatePassword')->name('update-password');
     Route::post('/update-profile-pic', 'UserController@updateProfilePic')->name('update-profile-pic');
 
     Route::get('/profile', 'UserController@read')->name('profile');